+++ R/module_filter_manager.R
@@ -37,7 +37,7 @@
 #'   }
 #' )
 #' if (interactive()) {
-#'   runApp(app)
+#'   shinyApp(app$ui, app$server)
 #' }
 #'
 #' @keywords internal