+++ R/teal_slices.R
@@ -1,14 +1,19 @@
 #' Filter settings for teal applications
 #'
-#' Filter settings for teal applications
+#' Specify initial filter states and filtering settings for a `teal` app.
+#'
+#' Produces a `teal_slices` object.
+#' The `teal_slice` components will specify filter states that will be active when the app starts.
+#' Attributes (created with the named arguments) will configure the way the app applies filters.
+#' See argument descriptions for details.
 #'
 #' @inheritParams teal.slice::teal_slices
 #'
-#' @param module_specific (`logical(1)`)\cr
-#'  - `TRUE` when filter panel should be module-specific. All modules can have different set
-#'   of filters specified - see `mapping` argument.
-#'  - `FALSE` when one filter panel needed to all modules. All filters will be shared
-#'    by all modules.
+#' @param module_specific optional (`logical(1)`)\cr
+#'  - `FALSE` (default) when one filter panel applied to all modules.
+#'  All filters will be shared by all modules.
+#'  - `TRUE` when filter panel module-specific.
+#'  Modules can have different set of filters specified - see `mapping` argument.
 #' @param mapping `r lifecycle::badge("experimental")` _This is a new feature. Do kindly share your opinions.\cr_
 #'  (`named list`)\cr
 #'  Specifies which filters will be active in which modules on app start.
@@ -24,6 +29,11 @@
 #'  Used for verifying snapshots uploaded from file. See `snapshot`.
 #'
 #' @param x (`list`) of lists to convert to `teal_slices`
+#'
+#' @return
+#' A `teal_slices` object.
+#'
+#' @seealso [`teal.slice::teal_slices`], [`teal.slice::teal_slice`], [`slices_store`]
 #'
 #' @examples
 #' filter <- teal_slices(
@@ -41,16 +51,16 @@
 #' )
 #'
 #' app <- teal::init(
+#'   data = list(iris = iris, mtcars = mtcars),
 #'   modules = list(
 #'     module("module1"),
 #'     module("module2")
 #'   ),
-#'   data = list(iris, mtcars),
 #'   filter = filter
 #' )
 #'
 #' if (interactive()) {
-#'   shiny::runApp(app)
+#'   shinyApp(app$ui, app$server)
 #' }
 #'
 #' @export